Move OIME feature into tubes app, delete oime app
Model names (OIMEProposal, OIMESolverRole, OIMEAttempt, OIMEComment) and
URL names are unchanged. OIME views/forms/admin are merged into the tubes
app. Tests split into tubes/tests.py (tube-specific) and tubes/tests_oime.py.
Pyright exclude pattern updated to cover tests_*.py files.
