+++ otisweb/urls.py
@@ -25,7 +25,6 @@ urlpatterns = [
     path(r"roster/", include("roster.urls")),
     path(r"rpg/", include("rpg.urls")),
     path(r"payments/", include("payments.urls")),
-    path(r"oime/", include("oime.urls")),
     path(r"suggestions/", include("suggestions.urls")),
     path(r"tubes/", include("tubes.urls")),
     # ------
