+++ tubes/templates/tubes/proposal_detail.html
