+++ tubes/tests_oime.py
@@ -7,7 +7,7 @@ from .models import OIMEAttempt, OIMEComment, OIMESolverRole
 
 
 # ---------------------------------------------------------------------------
-# Verified-gating tests
+# Verified-gating
 # ---------------------------------------------------------------------------
 
 
@@ -101,7 +101,7 @@ def test_role_can_be_changed(otis):
 
 
 # ---------------------------------------------------------------------------
-# Proposal creation
+# Proposal creation / editing
 # ---------------------------------------------------------------------------
 
 
@@ -183,7 +183,7 @@ def test_staff_can_update_any_proposal(otis):
 
 
 # ---------------------------------------------------------------------------
-# Casual solver: full access, no timer
+# Casual solver
 # ---------------------------------------------------------------------------
 
 
@@ -215,7 +215,7 @@ def test_casual_can_comment(otis):
 
 
 # ---------------------------------------------------------------------------
-# Serious solver: timer and attempt flow
+# Serious solver
 # ---------------------------------------------------------------------------
